🎨 Move golden path to renamed 'flows' directory
