+++ apps/flows/src/app/app-routing.module.ts
@@ -3,6 +3,11 @@ import { PreloadAllModules, RouterModule, Routes } from '@angular/router';
 
 import { HomeComponent } from './home/home.component';
 import { LongListFlowComponent } from './long-list-flow/long-list-flow.component';
+import { AccountComponent } from './transaction-details-flow/account/account.component';
+import { ForYouComponent } from './transaction-details-flow/for-you/for-you.component';
+import { TransactionDetailsFlowComponent } from './transaction-details-flow/transaction-details-flow.component';
+import { TransactionDetailsComponent } from './transaction-details-flow/transaction-details/transaction-details.component';
+import { TransactionsComponent } from './transaction-details-flow/transactions/transactions.component';
 
 const routes: Routes = [
   {
@@ -12,6 +17,30 @@ const routes: Routes = [
   {
     path: 'long-list-flow',
     component: LongListFlowComponent,
+  },
+  {
+    path: 'transaction-details-flow',
+    component: TransactionDetailsFlowComponent,
+    children: [
+      {
+        path: '',
+        redirectTo: 'for-you',
+        pathMatch: 'full',
+      },
+      {
+        path: 'for-you',
+        component: ForYouComponent,
+      },
+      {
+        path: 'account',
+        component: AccountComponent,
+      },
+      { path: 'transactions', component: TransactionsComponent },
+      {
+        path: 'transactions-details/:id',
+        component: TransactionDetailsComponent,
+      },
+    ],
   },
 ];
