+++ apps/flows/src/app/home/home.component.html
@@ -23,16 +23,20 @@
         </div>
         <div class="grid-item half-at-tablet-up">
           <article>
-            <kirby-card [hasPadding]="true" mode="highlighted" (click)="navigate('long-list-flow')">
+            <kirby-card
+              [hasPadding]="true"
+              mode="highlighted"
+              (click)="navigate('transaction-details-flow')"
+            >
               <kirby-avatar
                 imageSrc="https://picsum.photos/100?random=2"
                 size="md"
                 shadow="true"
               ></kirby-avatar>
-              <h2 class="card-title">Transaction List with many Items</h2>
+              <h2 class="card-title">Navigate to List Item detail</h2>
               <p>
-                Showcase of Kirby List with 600 items. This flow can be useful when evaluating the
-                performance of items and lists in combination.
+                Showcase of a standard flow for navigating from a landing page to a clickable list
+                item that shows a detail page.
               </p>
             </kirby-card>
           </article>
