+++ apps/flows/src/app/long-list-flow/list-item/list-item.component.ts
@@ -1,6 +1,6 @@
 import { Component, Input } from '@angular/core';
 
-import { Transaction } from '../transaction';
+import { Transaction } from '../../transaction';
 
 @Component({
   selector: 'list-item',
