+++ apps/flows/src/app/transaction-details-flow/for-you/for-you.component.scss
