+++ apps/flows/src/app/transaction-details-flow/for-you/for-you.component.ts
@@ -1,0 +1,8 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'kirbydesign-transactions-details-foryou',
+  templateUrl: './for-you.component.html',
+  styleUrls: ['./for-you.component.scss'],
+})
+export class ForYouComponent {}
