+++ apps/flows/src/app/transaction-details-flow/transaction-details-flow.component.html
@@ -1,6 +1,6 @@
 <kirby-app>
   <kirby-tab-bar>
-    <kirby-tab-button routerLink="foryou">
+    <kirby-tab-button routerLink="for-you">
       <kirby-icon name="person-outline"></kirby-icon>
       <kirby-icon name="person" selected-tab></kirby-icon>
       Til dig
