+++ apps/flows/src/app/transaction-details-flow/transaction-details/transaction-details.component.html
