+++ apps/flows/src/app/transaction-details-flow/transactions/transactions.component.scss
