--- apps/user-flows/src/app/app.component.html
@@ -1,3 +1,0 @@
-<kirby-app>
-  <kirby-router-outlet></kirby-router-outlet>
-</kirby-app>
