--- apps/user-flows/src/app/app.component.ts
@@ -1,9 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'kirbydesign-root',
-  templateUrl: './app.component.html',
-})
-export class AppComponent {
-  title = 'user-flows';
-}
