--- apps/user-flows/src/app/app.module.ts
@@ -1,27 +1,0 @@
-import { NgModule } from '@angular/core';
-import { BrowserModule } from '@angular/platform-browser';
-
-import { KirbyModule } from '@kirbydesign/designsystem';
-
-import { AppRoutingModule } from './app-routing.module';
-import { AppComponent } from './app.component';
-import { AccountComponent } from './transaction-details-flow/account/account.component';
-import { ForyouComponent } from './transaction-details-flow/foryou/foryou.component';
-import { TransactionDetailsFlowComponent } from './transaction-details-flow/transaction-details-flow.component';
-import { TransactionDetailsComponent } from './transaction-details-flow/transaction-details/transaction-details.component';
-import { TransactionsComponent } from './transaction-details-flow/transactions/transactions.component';
-
-@NgModule({
-  declarations: [
-    AppComponent,
-    TransactionDetailsFlowComponent,
-    ForyouComponent,
-    AccountComponent,
-    TransactionsComponent,
-    TransactionDetailsComponent,
-  ],
-  imports: [BrowserModule, AppRoutingModule, KirbyModule],
-  providers: [],
-  bootstrap: [AppComponent],
-})
-export class AppModule {}
