--- apps/user-flows/src/app/home/base-list.component.ts
@@ -1,727 +1,0 @@
-export class BaseListComponent {
-  public itemsFullList: any[] = [
-    {
-      id: 0,
-      title: 'Vestas Wind Systems has a very long name',
-      subTitle: '2000 pcs',
-      amount: '5.587.218.309 DKK',
-      detail: 225,
-      color: 'default',
-    },
-    {
-      id: 1,
-      title: 'Cypress Semiconductor Corporation',
-      subTitle: '1827 pcs',
-      amount: '76.980 DKK',
-      detail: -3,
-      color: 'light',
-    },
-    {
-      id: 2,
-      title: 'Ultragenyx Pharmaceutical Inc.',
-      subTitle: '787 pcs',
-      amount: '83.004 DKK',
-      detail: -115,
-      color: 'white',
-    },
-    {
-      id: 3,
-      title: 'Trans World Entertainment Corp.',
-      subTitle: '467 pcs',
-      amount: '60.963 DKK',
-      detail: 6,
-      color: 'light',
-    },
-    {
-      id: 4,
-      title: 'Astronics Corporation',
-      subTitle: '791 pcs',
-      amount: '33.830 DKK',
-      detail: -154,
-      color: 'white',
-    },
-    {
-      id: 5,
-      title: 'Riverview Bancorp Inc',
-      subTitle: '206 pcs',
-      amount: '60.775 DKK',
-      detail: 98,
-      color: 'light',
-    },
-    {
-      id: 6,
-      title: 'Haemonetics Corporation',
-      subTitle: '988 pcs',
-      amount: '61.196 DKK',
-      detail: 220,
-      color: 'white',
-    },
-    {
-      id: 7,
-      title: 'PJT Partners Inc.',
-      subTitle: '1706 pcs',
-      amount: '52.441 DKK',
-      detail: 129,
-      color: 'light',
-    },
-    {
-      id: 8,
-      title: 'Total S.A.',
-      subTitle: '827 pcs',
-      amount: '62.276 DKK',
-      detail: 53,
-      color: 'white',
-    },
-    {
-      id: 9,
-      title: 'Northrop Grumman Corporation',
-      subTitle: '443 pcs',
-      amount: '95.190 DKK',
-      detail: -135,
-      color: 'white',
-    },
-    {
-      id: 10,
-      title: 'Rudolph Technologies, Inc.',
-      subTitle: '1799 pcs',
-      amount: '18.442 DKK',
-      detail: 86,
-    },
-    {
-      id: 11,
-      title: 'Atlas Financial Holdings, Inc.',
-      subTitle: '129 pcs',
-      amount: '75.842 DKK',
-      detail: -43,
-    },
-    {
-      id: 12,
-      title: 'Five Star Senior Living Inc.',
-      subTitle: '36 pcs',
-      amount: '99.125 DKK',
-      detail: -144,
-    },
-    {
-      id: 13,
-      title: 'Microbot Medical Inc. ',
-      subTitle: '1426 pcs',
-      amount: '35.459 DKK',
-      detail: -78,
-    },
-    {
-      id: 14,
-      title: 'Nova Lifestyle, Inc',
-      subTitle: '1840 pcs',
-      amount: '26.466 DKK',
-      detail: -84,
-    },
-    {
-      id: 15,
-      title: 'Corning Incorporated',
-      subTitle: '854 pcs',
-      amount: '37.490 DKK',
-      detail: -94,
-    },
-    {
-      id: 16,
-      title: 'Chembio Diagnostics, Inc.',
-      subTitle: '1964 pcs',
-      amount: '41.765 DKK',
-      detail: 122,
-    },
-    {
-      id: 17,
-      title: 'RLJ Lodging Trust',
-      subTitle: '393 pcs',
-      amount: '68.623 DKK',
-      detail: 14,
-    },
-    {
-      id: 18,
-      title: 'Prudential Financial, Inc.',
-      subTitle: '171 pcs',
-      amount: '45.999 DKK',
-      detail: -77,
-    },
-    {
-      id: 19,
-      title: 'PDC Energy, Inc.',
-      subTitle: '1837 pcs',
-      amount: '52.101 DKK',
-      detail: -152,
-    },
-    {
-      id: 20,
-      title: 'Norbord Inc.',
-      subTitle: '1271 pcs',
-      amount: '36.160 DKK',
-      detail: 183,
-    },
-    {
-      id: 21,
-      title: 'Origo Acquisition Corporation',
-      subTitle: '929 pcs',
-      amount: '49.608 DKK',
-      detail: 174,
-    },
-    {
-      id: 22,
-      title: 'American Railcar Industries, Inc.',
-      subTitle: '861 pcs',
-      amount: '10.419 DKK',
-      detail: -179,
-    },
-    {
-      id: 23,
-      title: 'QAD Inc.',
-      subTitle: '1195 pcs',
-      amount: '53.019 DKK',
-      detail: -156,
-    },
-    {
-      id: 24,
-      title: 'ENDRA Life Sciences Inc.',
-      subTitle: '1522 pcs',
-      amount: '36.737 DKK',
-      detail: -147,
-    },
-    {
-      id: 25,
-      title: 'Pacific Mercantile Bancorp',
-      subTitle: '958 pcs',
-      amount: '95.228 DKK',
-      detail: 249,
-    },
-    {
-      id: 26,
-      title: 'Aerojet Rocketdyne Holdings, Inc. ',
-      subTitle: '27 pcs',
-      amount: '92.171 DKK',
-      detail: 82,
-    },
-    {
-      id: 27,
-      title: 'Simon Property Group, Inc.',
-      subTitle: '281 pcs',
-      amount: '5.875 DKK',
-      detail: -121,
-    },
-    {
-      id: 28,
-      title: "Dr. Reddy's Laboratories Ltd",
-      subTitle: '1882 pcs',
-      amount: '5.498 DKK',
-      detail: 66,
-    },
-    {
-      id: 29,
-      title: 'Everest Re Group, Ltd.',
-      subTitle: '941 pcs',
-      amount: '24.855 DKK',
-      detail: -1,
-    },
-    {
-      id: 30,
-      title: 'Gores Holdings II, Inc.',
-      subTitle: '1689 pcs',
-      amount: '86.716 DKK',
-      detail: -150,
-    },
-    {
-      id: 31,
-      title: 'Bank Mutual Corporation',
-      subTitle: '1568 pcs',
-      amount: '25.747 DKK',
-      detail: -197,
-    },
-    {
-      id: 32,
-      title: 'Two Harbors Investments Corp',
-      subTitle: '936 pcs',
-      amount: '64.032 DKK',
-      detail: -103,
-    },
-    {
-      id: 33,
-      title: 'Chesapeake Energy Corporation',
-      subTitle: '287 pcs',
-      amount: '51.267 DKK',
-      detail: 5,
-    },
-    {
-      id: 34,
-      title: 'Royce Global Value Trust, Inc.',
-      subTitle: '656 pcs',
-      amount: '45.622 DKK',
-      detail: -178,
-    },
-    {
-      id: 35,
-      title: 'R.R. Donnelley & Sons Company',
-      subTitle: '587 pcs',
-      amount: '65.442 DKK',
-      detail: -117,
-    },
-    {
-      id: 36,
-      title: 'First Bank',
-      subTitle: '492 pcs',
-      amount: '19.560 DKK',
-      detail: 48,
-    },
-    {
-      id: 37,
-      title: 'Digital Realty Trust, Inc.',
-      subTitle: '185 pcs',
-      amount: '60.367 DKK',
-      detail: 169,
-    },
-    {
-      id: 38,
-      title: 'Kenon Holdings Ltd.',
-      subTitle: '907 pcs',
-      amount: '80.383 DKK',
-      detail: -75,
-    },
-    {
-      id: 39,
-      title: 'Pebblebrook Hotel Trust',
-      subTitle: '672 pcs',
-      amount: '93.421 DKK',
-      detail: 32,
-    },
-    {
-      id: 40,
-      title: 'Cable One, Inc.',
-      subTitle: '329 pcs',
-      amount: '67.225 DKK',
-      detail: 249,
-    },
-    {
-      id: 41,
-      title: 'Acacia Research Corporation',
-      subTitle: '1060 pcs',
-      amount: '4.490 DKK',
-      detail: 196,
-    },
-    {
-      id: 42,
-      title: 'Ashford Hospitality Trust Inc',
-      subTitle: '1926 pcs',
-      amount: '31.059 DKK',
-      detail: 213,
-    },
-    {
-      id: 43,
-      title: 'MakeMyTrip Limited',
-      subTitle: '1177 pcs',
-      amount: '59.227 DKK',
-      detail: 110,
-    },
-    {
-      id: 44,
-      title: 'Icahn Enterprises L.P.',
-      subTitle: '663 pcs',
-      amount: '10.415 DKK',
-      detail: 99,
-    },
-    {
-      id: 45,
-      title: 'China Unicom (Hong Kong) Ltd',
-      subTitle: '517 pcs',
-      amount: '30.143 DKK',
-      detail: 90,
-    },
-    {
-      id: 46,
-      title: 'Intrexon Corporation',
-      subTitle: '1739 pcs',
-      amount: '4.334 DKK',
-      detail: -70,
-    },
-    {
-      id: 47,
-      title: 'Sequans Communications S.A.',
-      subTitle: '109 pcs',
-      amount: '46.363 DKK',
-      detail: 160,
-    },
-    {
-      id: 48,
-      title: 'EMC Insurance Group Inc.',
-      subTitle: '219 pcs',
-      amount: '73.717 DKK',
-      detail: -40,
-    },
-    {
-      id: 49,
-      title: 'Corbus Pharmaceuticals Holdings, Inc.',
-      subTitle: '1610 pcs',
-      amount: '86.740 DKK',
-      detail: 74,
-    },
-    {
-      id: 50,
-      title: 'China Ceramics Co., Ltd.',
-      subTitle: '1656 pcs',
-      amount: '68.075 DKK',
-      detail: 237,
-    },
-    {
-      id: 51,
-      title: 'Chemung Financial Corp',
-      subTitle: '1044 pcs',
-      amount: '20.552 DKK',
-      detail: 31,
-    },
-    {
-      id: 52,
-      title: 'Connecture, Inc.',
-      subTitle: '62 pcs',
-      amount: '40.611 DKK',
-      detail: 10,
-    },
-    {
-      id: 53,
-      title: 'PROS Holdings, Inc.',
-      subTitle: '377 pcs',
-      amount: '92.387 DKK',
-      detail: 104,
-    },
-    {
-      id: 54,
-      title: 'First Financial Bancorp.',
-      subTitle: '681 pcs',
-      amount: '81.602 DKK',
-      detail: -22,
-    },
-    {
-      id: 55,
-      title: 'iShares MSCI EM ESG Optimized ETF',
-      subTitle: '1196 pcs',
-      amount: '80.588 DKK',
-      detail: -62,
-    },
-    {
-      id: 56,
-      title: 'Western Asset Managed Municipals Fund, Inc.',
-      subTitle: '1939 pcs',
-      amount: '77.171 DKK',
-      detail: -182,
-    },
-    {
-      id: 57,
-      title: 'CTI Industries Corporation',
-      subTitle: '1657 pcs',
-      amount: '46.813 DKK',
-      detail: -90,
-    },
-    {
-      id: 58,
-      title: 'Paycom Software, Inc.',
-      subTitle: '1609 pcs',
-      amount: '58.124 DKK',
-      detail: -141,
-    },
-    {
-      id: 59,
-      title: 'Tower International, Inc.',
-      subTitle: '1157 pcs',
-      amount: '6.468 DKK',
-      detail: 224,
-    },
-    {
-      id: 60,
-      title: 'Sarepta Therapeutics, Inc.',
-      subTitle: '963 pcs',
-      amount: '50.810 DKK',
-      detail: 103,
-    },
-    {
-      id: 61,
-      title: 'Golar LNG Partners LP',
-      subTitle: '13 pcs',
-      amount: '47.434 DKK',
-      detail: 30,
-    },
-    {
-      id: 62,
-      title: 'Applied Genetic Technologies Corporation',
-      subTitle: '1022 pcs',
-      amount: '10.281 DKK',
-      detail: 30,
-    },
-    {
-      id: 63,
-      title: 'CHS Inc',
-      subTitle: '1260 pcs',
-      amount: '81.019 DKK',
-      detail: 212,
-    },
-    {
-      id: 64,
-      title: 'Principal Real Estate Income Fund',
-      subTitle: '408 pcs',
-      amount: '99.185 DKK',
-      detail: 69,
-    },
-    {
-      id: 65,
-      title: 'Nuveen Real Estate Fund',
-      subTitle: '1107 pcs',
-      amount: '29.376 DKK',
-      detail: -74,
-    },
-    {
-      id: 66,
-      title: 'Oaktree Capital Group, LLC',
-      subTitle: '840 pcs',
-      amount: '75.339 DKK',
-      detail: 248,
-    },
-    {
-      id: 67,
-      title: 'Fresh Del Monte Produce, Inc.',
-      subTitle: '957 pcs',
-      amount: '27.564 DKK',
-      detail: -150,
-    },
-    {
-      id: 68,
-      title: 'MVC Capital, Inc.',
-      subTitle: '281 pcs',
-      amount: '69.197 DKK',
-      detail: 47,
-    },
-    {
-      id: 69,
-      title: 'MercadoLibre, Inc.',
-      subTitle: '1113 pcs',
-      amount: '57.923 DKK',
-      detail: 203,
-    },
-    {
-      id: 70,
-      title: 'SmartFinancial, Inc.',
-      subTitle: '735 pcs',
-      amount: '89.855 DKK',
-      detail: 146,
-    },
-    {
-      id: 71,
-      title: 'ClubCorp Holdings, Inc.',
-      subTitle: '313 pcs',
-      amount: '99.364 DKK',
-      detail: -86,
-    },
-    {
-      id: 72,
-      title: 'Immunomedics, Inc.',
-      subTitle: '71 pcs',
-      amount: '69.832 DKK',
-      detail: 188,
-    },
-    {
-      id: 73,
-      title: 'KapStone Paper and Packaging Corporation',
-      subTitle: '473 pcs',
-      amount: '12.219 DKK',
-      detail: -64,
-    },
-    {
-      id: 74,
-      title: 'Stifel Financial Corporation',
-      subTitle: '564 pcs',
-      amount: '58.651 DKK',
-      detail: 154,
-    },
-    {
-      id: 75,
-      title: 'Xinyuan Real Estate Co Ltd',
-      subTitle: '810 pcs',
-      amount: '72.391 DKK',
-      detail: 198,
-    },
-    {
-      id: 76,
-      title: 'Liberty Global plc',
-      subTitle: '1473 pcs',
-      amount: '1.976 DKK',
-      detail: -135,
-    },
-    {
-      id: 77,
-      title: 'Morgan Stanley',
-      subTitle: '1079 pcs',
-      amount: '58.112 DKK',
-      detail: 41,
-    },
-    {
-      id: 78,
-      title: 'Evertec, Inc.',
-      subTitle: '1851 pcs',
-      amount: '59.546 DKK',
-      detail: 48,
-    },
-    {
-      id: 79,
-      title: 'CVR Refining, LP',
-      subTitle: '1134 pcs',
-      amount: '99.919 DKK',
-      detail: 91,
-    },
-    {
-      id: 80,
-      title: 'Marine Petroleum Trust',
-      subTitle: '228 pcs',
-      amount: '1.030 DKK',
-      detail: 240,
-    },
-    {
-      id: 81,
-      title: 'Frontier Communications Corporation',
-      subTitle: '1534 pcs',
-      amount: '66.856 DKK',
-      detail: -70,
-    },
-    {
-      id: 82,
-      title: 'Middlesex Water Company',
-      subTitle: '1584 pcs',
-      amount: '27.582 DKK',
-      detail: 109,
-    },
-    {
-      id: 83,
-      title: 'Syneron Medical Ltd.',
-      subTitle: '157 pcs',
-      amount: '76.750 DKK',
-      detail: 73,
-    },
-    {
-      id: 84,
-      title: 'Allison Transmission Holdings, Inc.',
-      subTitle: '857 pcs',
-      amount: '98.364 DKK',
-      detail: -20,
-    },
-    {
-      id: 85,
-      title: 'Cirrus Logic, Inc.',
-      subTitle: '95 pcs',
-      amount: '79.359 DKK',
-      detail: -102,
-    },
-    {
-      id: 86,
-      title: 'MFS Multimarket Income Trust',
-      subTitle: '687 pcs',
-      amount: '17.861 DKK',
-      detail: -95,
-    },
-    {
-      id: 87,
-      title: 'Lincoln Electric Holdings, Inc.',
-      subTitle: '441 pcs',
-      amount: '71.454 DKK',
-      detail: 241,
-    },
-    {
-      id: 88,
-      title: 'Navient Corporation',
-      subTitle: '1034 pcs',
-      amount: '77.255 DKK',
-      detail: 30,
-    },
-    {
-      id: 89,
-      title: 'The Descartes Systems Group Inc.',
-      subTitle: '1213 pcs',
-      amount: '99.418 DKK',
-      detail: 12,
-    },
-    {
-      id: 90,
-      title: 'Republic First Bancorp, Inc.',
-      subTitle: '1366 pcs',
-      amount: '24.934 DKK',
-      detail: -46,
-    },
-    {
-      id: 91,
-      title: 'PNM Resources, Inc. (Holding Co.)',
-      subTitle: '79 pcs',
-      amount: '9.423 DKK',
-      detail: 32,
-    },
-    {
-      id: 92,
-      title: 'Arthur J. Gallagher & Co.',
-      subTitle: '299 pcs',
-      amount: '87.935 DKK',
-      detail: -105,
-    },
-    {
-      id: 93,
-      title: 'Rambus, Inc.',
-      subTitle: '564 pcs',
-      amount: '76.053 DKK',
-      detail: -129,
-    },
-    {
-      id: 94,
-      title: 'FARO Technologies, Inc.',
-      subTitle: '1912 pcs',
-      amount: '48.066 DKK',
-      detail: -141,
-    },
-    {
-      id: 95,
-      title: 'Dynex Capital, Inc.',
-      subTitle: '557 pcs',
-      amount: '81.093 DKK',
-      detail: -76,
-    },
-    {
-      id: 96,
-      title: 'Nuveen New York Municipal Value Fund, Inc.',
-      subTitle: '106 pcs',
-      amount: '88.488 DKK',
-      detail: -16,
-    },
-    {
-      id: 97,
-      title: 'Kirby Corporation',
-      subTitle: '406 pcs',
-      amount: '2.622 DKK',
-      detail: 217,
-    },
-    {
-      id: 98,
-      title: 'Oconee Federal Financial Corp.',
-      subTitle: '1649 pcs',
-      amount: '58.086 DKK',
-      detail: 102,
-    },
-    {
-      id: 99,
-      title: 'WSFS Financial Corporation',
-      subTitle: '1959 pcs',
-      amount: '28.587 DKK',
-      detail: 52,
-    },
-    {
-      id: 100,
-      title: 'Nuveen Short Duration Credit Opportunities Fund',
-      subTitle: '1835 pcs',
-      amount: '88.272 DKK',
-      detail: 75,
-    },
-  ];
-
-  public items: any[] = this.itemsFullList.slice(0, 8);
-
-  onItemSelect(item: any) {
-    alert(`You have clicked the row [${item.title} ${item.amount}]`);
-  }
-}
