--- apps/user-flows/src/app/home/home-routing.module.ts
@@ -1,17 +1,0 @@
-import { NgModule } from '@angular/core';
-import { RouterModule, Routes } from '@angular/router';
-
-import { HomePageComponent } from './home.page';
-
-const routes: Routes = [
-  {
-    path: '',
-    component: HomePageComponent,
-  },
-];
-
-@NgModule({
-  imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule],
-})
-export class HomePageRoutingModule {}
