--- apps/user-flows/src/app/home/transaction-item/transaction-item.component.ts
@@ -1,11 +1,0 @@
-import { Component, Input } from '@angular/core';
-
-import { Transaction } from '../../transaction';
-
-@Component({
-  selector: 'transaction-item',
-  templateUrl: './transaction-item.component.html',
-})
-export class TransactionItemComponent {
-  @Input() transaction: Transaction;
-}
