--- apps/user-flows/src/app/transaction-details-flow/foryou/foryou.component.ts
@@ -1,12 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'kirbydesign-transactions-details-foryou',
-  templateUrl: './foryou.component.html',
-  styleUrls: ['./foryou.component.scss'],
-})
-export class ForyouComponent implements OnInit {
-  constructor() {}
-
-  ngOnInit(): void {}
-}
