--- apps/user-flows/src/styles.scss
@@ -1,4 +1,0 @@
-/* You can add global styles to this file, and also import other style files */
-
-@use '~@kirbydesign/core/src/scss/global-styles';
-@use '~@kirbydesign/core/src/scss/utils';
