--- apps/user-flows/tsconfig.editor.json
@@ -1,7 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "include": ["**/*.ts"],
-  "compilerOptions": {
-    "types": ["jasmine", "node"]
-  }
-}
