--- apps/user-flows/tsconfig.json
@@ -1,37 +1,0 @@
-{
-  "extends": "../../tsconfig.base.json",
-  "files": [],
-  "include": [],
-  "references": [
-    {
-      "path": "./tsconfig.app.json"
-    },
-    {
-      "path": "./tsconfig.spec.json"
-    },
-    {
-      "path": "./tsconfig.editor.json"
-    }
-  ],
-  "compilerOptions": {
-    "forceConsistentCasingInFileNames": true,
-    "strict": false,
-    "noImplicitOverride": false,
-    "noPropertyAccessFromIndexSignature": false,
-    "noImplicitReturns": false,
-    "noFallthroughCasesInSwitch": false,
-    "resolveJsonModule": true,
-    "allowSyntheticDefaultImports": true,
-    "paths": {
-      "~/app/*": ["apps/user-flows/src/app/*"],
-      "~/environments/*": ["apps/user-flows/src/environments/*"],
-      "@kirbydesign/designsystem": ["libs/designsystem/src/lib/index.ts"],
-      "@kirbydesign/designsystem/*": ["libs/designsystem/src/lib/*"]
-    }
-  },
-  "angularCompilerOptions": {
-    "strictInjectionParameters": false,
-    "strictInputAccessModifiers": false,
-    "strictTemplates": false
-  }
-}
