+++ package.json
@@ -81,11 +81,11 @@
     "test:core:single": "stencil test --config=./libs/core/stencil.config.ts --spec",
     "workspace-generator": "nx workspace-generator",
     "sync-node-version": "node ./scripts/sync-node-version.js",
-    "start:user-flows:prod": "ng serve --configuration production",
-    "build:user-flows": "ng build user-flows",
-    "build:user-flows:prod": "npm run build:user-flows -- --configuration production",
-    "build:user-flows:prod:stats": "npm run build:user-flows:prod -- --stats-json && npm run copy-bundle-stats",
-    "analyze:user-flows:bundle": "npx webpack-bundle-analyzer ./measures/stats.json",
+    "start:flows:prod": "ng serve flows --configuration production",
+    "build:flows": "ng build flows",
+    "build:flows:prod": "npm run build:flows -- --configuration production",
+    "build:flows:prod:stats": "npm run build:flows:prod -- --stats-json && npm run copy-bundle-stats",
+    "analyze:flows:bundle": "npx webpack-bundle-analyzer ./measures/stats.json",
     "copy-bundle-stats": "copyfiles -f ./dist/apps/user-flows/stats.json measures"
   },
   "private": true,
